Energy is now exclusive to HumanPlayers. I may eventually do away with the energy system alltogether, as it slows down gameplay quite a bit. Next, on to DamageBacklog.
